Move parser to commands namespace
