+++ Unit4.Automation.Tests/Parser/ConfigCommandParserTests.cs
@@ -7,6 +7,7 @@ using Criteria = Unit4.Automation.Tests.Helpers.A.Criteria;
 using Unit4.Automation.Tests.Helpers;
 using System.Collections.Generic;
 using System.Text;
+using Unit4.Automation.Commands;
 
 namespace Unit4.Automation.Tests.Parser
 {
