+++ Unit4/Program.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using Unit4.Automation.Model;
+using Unit4.Automation.Commands;
 
 namespace Unit4.Automation
 {
